TS: fix socket/transport types
Fixes #743.

* Create Socket type definitions.
* Fix ondisconnect definition for Socket (previously defined in
  WebsocketInterface).
* Create missing definitions for UA connection events.
