+++ lib/UA.d.ts
@@ -1,6 +1,6 @@
 import {EventEmitter, Listener} from 'events'
 
-import {DisconnectEvent, Socket, WeightedSocket} from './WebSocketInterface'
+import {Socket, WeightedSocket} from './Socket'
 import {AnswerOptions, Originator, RTCSession, RTCSessionEventMap, TerminateOptions} from './RTCSession'
 import {IncomingRequest, IncomingResponse, OutgoingRequest} from './SIPMessage'
 import {Message, SendMessageOptions} from './Message'
@@ -60,13 +60,20 @@ export interface OutgoingRTCSessionEvent {
 
 export type RTCSessionEvent = IncomingRTCSessionEvent | OutgoingRTCSessionEvent;
 
-export interface UAConnectingEvent {
+export interface ConnectingEvent {
   socket: Socket;
   attempts: number
 }
 
 export interface ConnectedEvent {
   socket: Socket;
+}
+
+export interface DisconnectEvent {
+  socket: Socket;
+  error: boolean;
+  code?: number;
+  reason?: string;
 }
 
 export interface RegisteredEvent {
@@ -100,7 +107,7 @@ export interface OutgoingOptionsEvent {
   request: OutgoingRequest;
 }
 
-export type UAConnectingListener = (event: UAConnectingEvent) => void;
+export type ConnectingListener = (event: ConnectingEvent) => void;
 export type ConnectedListener = (event: ConnectedEvent) => void;
 export type DisconnectedListener = (event: DisconnectEvent) => void;
 export type RegisteredListener = (event: RegisteredEvent) => void;
@@ -118,7 +125,7 @@ export type OptionsListener = IncomingOptionsListener | OutgoingOptionsListener;
 export type SipEventListener = <T = any>(event: { event: T; request: IncomingRequest; }) => void
 
 export interface UAEventMap {
-  connecting: UAConnectingListener;
+  connecting: ConnectingListener;
   connected: ConnectedListener;
   disconnected: DisconnectedListener;
   registered: RegisteredListener;
